React, Spring, MySQL

+++ src/main/java/com/fullstackproject/web/CommentAndLikesController.java
@@ -9,9 +9,11 @@ import com.fullstackproject.repositories.CommentRepository;
 import com.fullstackproject.repositories.JokeRepository;
 import com.fullstackproject.repositories.LikeRepository;
 import com.fullstackproject.repositories.UserRepository;
+import com.fullstackproject.security.rolesAuth.IsProfileUser;
 import org.modelmapper.ModelMapper;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
@@ -49,10 +51,12 @@ public class CommentAndLikesController {
     @PreAuthorize("isAuthenticated()")
     public Object addCommentToJoke(@PathVariable String jokeId,
                                    @RequestBody @Valid CommentDto commentDto, BindingResult bindingResult) {
+        ErrorRest errorRest = new ErrorRest();
+        ErrorRest checkForAuth = getError(commentDto, errorRest);
+        if (checkForAuth != null) return checkForAuth;
 
         if (bindingResult.hasErrors()) {
 
-            ErrorRest errorRest = new ErrorRest();
             List<FieldError> errors = bindingResult.getFieldErrors();
             List<String> message = new ArrayList<>();
             for (FieldError e : errors) {
@@ -79,8 +83,12 @@ public class CommentAndLikesController {
     @DeleteMapping("/add/comment/:{id}")
     @PreAuthorize("isAuthenticated()")
     public ResponseEntity<?> deleteCommentById(@PathVariable String id) {
-
         Comment comment = this.commentRepository.findById(id).get();
+
+        String principal = SecurityContextHolder.getContext().getAuthentication().getName();
+        if (!principal.equals(comment.getOwnerOfComment())) {
+            return ResponseEntity.status(401).build();
+        }
 
         this.commentRepository.deleteById(id);
 
@@ -89,6 +97,7 @@ public class CommentAndLikesController {
 
     @PostMapping("/add/like/:{id}/:{username}")
     @PreAuthorize("isAuthenticated()")
+    @IsProfileUser
     public ResponseEntity<?> addLike(@PathVariable String id, @PathVariable String username) {
 
         Joke joke = this.jokeRepository.findById(id).get();
@@ -99,4 +108,17 @@ public class CommentAndLikesController {
         this.likeRepository.save(like);
         return ResponseEntity.status(200).build();
     }
+
+    private ErrorRest getError(CommentDto commentDto, ErrorRest errorRest) {
+        String principal = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        if (!principal.equals(commentDto.getOwnerOfComment())) {
+            errorRest.setCode(400);
+            errorRest.setMessage("Adding comment Failed");
+            errorRest.setCause("Failed adding!");
+            return errorRest;
+        }
+        return null;
+    }
+
 }
